requirements.txt
flask
gunicorn
mysql-connector-python
bs4
requests
datetime
flask_cors
selenium
webdriver-manager
smtplib
ssl